requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
jupyter==1.0.0
notebook==6.5.6
ipykernel==6.25.2